ref: Change criterion for 'false' ScatterInfo
Before we compared with the ScatterInfo::miss() singleton, but we now need
the ability to create various 'false' ScatterInfo s, as a 'false' ScatterInfo
still might contain valid emissive values, now that we want to introduce lights.
For that reason the ScatterInfo is already declared false when the ray direction
is zero. This allows to still use the remaining fields of ScatterInfo.