ref: move scene stuff from main.cpp to own file
I wanted to thin out the main file to make it easier to add user input handling
to the main file. This in turn is necessary to make it easier to investigate the
parallelization behavior.
All this scene stuff kind of belongs together, so I might as well put it into
a separate file.